+++ src/inc/api/APISendProgress.class.php
@@ -12,6 +12,7 @@ use DBA\QueryFilter;
 use DBA\Zap;
 use DBA\QueryFilterWithNull;
 use DBA\TaskDebugOutput;
+use DBA\AgentStat;
 
 class APISendProgress extends APIBasic {
   public function execute($QUERY = array()) {
@@ -63,6 +64,28 @@ class APISendProgress extends APIBasic {
       if ($hashlist->getIsSecret() > $this->agent->getIsTrusted()) {
         $this->sendErrorResponse(PActions::SEND_PROGRESS, "Unknown Error. The API does not trust you with more information");
       }
+    }
+
+    $dataTime = time();
+    if(isset($QUERY[PQuerySendProgress::GPU_TEMP])){
+      for($i =0; $i < sizeof($QUERY[PQuerySendProgress::GPU_TEMP]); $i++){
+        if(!is_numeric($QUERY[PQuerySendProgress::GPU_TEMP][$i]) || $QUERY[PQuerySendProgress::GPU_TEMP][$i] <= 0){
+          unset($QUERY[PQuerySendProgress::GPU_TEMP][$i]);
+        }
+      }
+      $data = implode(",", $QUERY[PQuerySendProgress::GPU_TEMP]);
+      $agentStat = new AgentStat(0, $this->agent->getId(), DAgentStatsType::GPU_TEMP, $dataTime, $data);
+      $FACTORIES::getAgentStatFactory()->save($agentStat);
+    }
+    if(isset($QUERY[PQuerySendProgress::GPU_UTIL])){
+      for($i =0; $i < sizeof($QUERY[PQuerySendProgress::GPU_UTIL]); $i++){
+        if(!is_numeric($QUERY[PQuerySendProgress::GPU_UTIL][$i]) || $QUERY[PQuerySendProgress::GPU_UTIL][$i] < 0){
+          unset($QUERY[PQuerySendProgress::GPU_UTIL][$i]);
+        }
+      }
+      $data = implode(",", $QUERY[PQuerySendProgress::GPU_UTIL]);
+      $agentStat = new AgentStat(0, $this->agent->getId(), DAgentStatsType::GPU_UTIL, $dataTime, $data);
+      $FACTORIES::getAgentStatFactory()->save($agentStat);
     }
 
     // agent is assigned to this chunk (not necessarily task!)
@@ -388,6 +411,7 @@ class APISendProgress extends APIBasic {
         break;
     }
     Util::zapCleaning();
+    Util::agentStatCleaning();
     $this->sendResponse(array(
         PResponseSendProgress::ACTION => PActions::SEND_PROGRESS,
         PResponseSendProgress::RESPONSE => PValues::SUCCESS,