+++ src/inc/defines/config.php
@@ -38,6 +38,7 @@ class DConfig {
   const RULE_SPLIT_ALWAYS      = "ruleSplitAlways";
   const RULE_SPLIT_DISABLE     = "ruleSplitDisable";
   const PRINCE_LINK            = "princeLink";
+  const AGENT_DATA_LIFETIME    = "agentDataLifetime";
 
   // Section: Yubikey
   const YUBIKEY_ID  = "yubikey_id";
@@ -59,6 +60,8 @@ class DConfig {
   const HASHES_PER_PAGE       = "hashesPerPage";
   const HIDE_IP_INFO          = "hideIpInfo";
   const SHOW_TASK_PERFORMANCE = "showTaskPerformance";
+  const AGENT_STAT_LIMIT      = "agentStatLimit";
+  const AGENT_STAT_TENSION    = "agentStatTension";
 
   // Section: Server
   const BASE_URL           = "baseUrl";
@@ -144,6 +147,12 @@ class DConfig {
         return DConfigType::TICKBOX;
       case DConfig::PRINCE_LINK:
         return DConfigType::STRING_INPUT;
+      case DConfig::AGENT_STAT_LIMIT:
+        return DConfigType::NUMBER_INPUT;
+      case DConfig::AGENT_DATA_LIFETIME:
+        return DConfigType::NUMBER_INPUT;
+      case DConfig::AGENT_STAT_TENSION:
+        return DConfigType::TICKBOX;
     }
     return DConfigType::STRING_INPUT;
   }
@@ -226,6 +235,12 @@ class DConfig {
         return "Disable automatic task splitting with large rule files.";
       case DConfig::PRINCE_LINK:
         return "Download link for the prince preprocessor binaries.";
+      case DConfig::AGENT_STAT_LIMIT:
+        return "Maximal number of data points showing of agent gpu data.";
+      case DConfig::AGENT_DATA_LIFETIME:
+        return "Minimum time in seconds how long agent data about gpu temperature and utility is kept on the server.";
+      case DConfig::AGENT_STAT_TENSION:
+        return "Draw straigth lines in agent data graph  instead of bezier curves.";
     }
     return $config;
   }