+++ src/install/updates/update_v0.7.0_v0.x.x.php
@@ -2,6 +2,7 @@
 
 use DBA\Factory;
 use DBA\File;
+use DBA\Config;
 
 require_once(dirname(__FILE__) . "/../../inc/db.php");
 require_once(dirname(__FILE__) . "/../../dba/init.php");
@@ -33,6 +34,20 @@ echo "Add file access groups... ";
 $FACTORIES::getAgentFactory()->getDB()->query("ALTER TABLE `File` ADD `accessGroupId` INT NOT NULL");
 $uS = new UpdateSet(File::ACCESS_GROUP_ID, AccessUtils::getOrCreateDefaultAccessGroup()->getId());
 $FACTORIES::getFileFactory()->massUpdate(array($FACTORIES::UPDATE => $uS));
+echo "OK\n";
+
+echo "Update agent stats table... ";
+$FACTORIES::getAgentFactory()->getDB()->query("ALTER TABLE `AgentStat` CHANGE `time` `time` BIGINT(11) NOT NULL");
+$FACTORIES::getAgentFactory()->getDB()->query("ALTER TABLE `AgentStat` CHANGE `value` `value` VARCHAR(64) NOT NULL");
+echo "OK\n";
+
+echo "Add new config entries... ";
+$config = new Config(0, 4, DConfig::AGENT_STAT_LIMIT, 100);
+$FACTORIES::getConfigFactory()->save($config);
+$config = new Config(0, 1, DConfig::AGENT_DATA_LIFETIME, 3600);
+$FACTORIES::getConfigFactory()->save($config);
+$config = new Config(0, 4, DConfig::AGENT_STAT_TENSION, 0);
+$FACTORIES::getConfigFactory()->save($config);
 echo "OK\n";
 
 echo "Update complete!\n";