+++ doc/changelog.md
@@ -4,6 +4,7 @@
 
 - The server can store sent debug output from Hashcat sent by the agent.
 - Files now also are associated to an Access Group to control the visibility of files.
+- Agent data about device temperature and util is collected and can be viewed on the server.
 
 ## Bugfixes